% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_addtimes.R
\name{add_times}
\alias{add_times}
\title{Modify the quiz_object to add the order of answering
and the spent time per question}
\usage{
add_times(course)
}
\arguments{
\item{file}{a quiz object from read_lc}
}
\value{
it returns a quiz object containing the following columns:  
"email address", "question", "responded at", "score" and "quiz"
}
\description{
This function receives as input a quiz object 
created with the read_lc() function
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)
quiz_object <- add_times(quiz_object)
}
