\name{easiness_time}
\alias{easiness_time}
\title{Easiness-time plot}
\description{
  This function receives as input a catalytics file (excel of csv) and parse it to create a long format data.frame
}
\usage{
  easiness_time(quiz_dir)
  This function only receives as input the directory where all the quizzes files are stored.
}
\examples{
  item3
}
