\name{order_plot}
\alias{odder_plot}
\title{Order plot with ggplot2}
\description{
  This function computes a graph with the order in which the students answer the quizzes
}
\usage{
  order_plot(quiz_dir)
  This function only receives as input the directory where all the quizzes files are stored.
}
\examples{
  %item2
}
