\name{read_lc}
\alias{read_lc}
\title{Read catalytics file}
\description{
This function receives as input a catalytics file (excel of csv) and parse it to create a long format data.frame
}
\usage{
  read_lc(quiz_dir)
  This function only receives as input the directory where all the quizzes files are stored.
}
\examples{
  a <- 1:100
  a
}
