% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_boxplot.R
\name{plot_boxplots}
\alias{plot_boxplots}
\title{boxplot of the quiz with ggplot2}
\usage{
plot_boxplots(quiz_object)
}
\arguments{
\item{quiz_object}{Receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the boxplot
}
\description{
This function computes an boxplot of the grades per quiz
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)
plot_boxplots(quiz_object)
}
