% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_easiness_time.R
\name{plot_easiness_time}
\alias{plot_easiness_time}
\title{Easiness-Time plot.}
\usage{
plot_easiness_time(quiz_object)
}
\arguments{
\item{df_quizzes}{receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the easiness-time of the selected quiz
}
\description{
This function computes a graph displaying the easiness-time plot for each quiz.
}
\examples{
file_to_read <- "../../datasets/Dataset1/Quiz3_session12098.csv"
quiz_object <- read_lc(file_to_read)
quiz_object <- add_times(quiz_object)
plot_easiness_time(quiz_object)
}
