% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_guessers.R
\name{plot_guessers}
\alias{plot_guessers}
\title{Guessers per quiz plot}
\usage{
plot_guessers(df_quizzes)
}
\arguments{
\item{df_quizzes}{receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the guessers of the selected quiz
}
\description{
This function computes a graph displaying the guessers plot for each quiz.
}
\examples{
file_to_read <- "../../datasets/Dataset1/Quiz3_session12098.csv"
long_format <- read_lc(file_to_read)
plot_guessers(long_format)
}
