% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_histogram.R
\name{plot_hist}
\alias{plot_hist}
\title{histogram of the quiz with ggplot2}
\usage{
plot_hist(quiz_object)
}
\arguments{
\item{quiz_object}{Receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the histogram
}
\description{
This function computes an histogram of the grades
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)
plot_hist(quiz_object)
}
