% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_ICC.R
\name{plot_jointICC}
\alias{plot_jointICC}
\title{Item Characteristic Curve ICC}
\usage{
plot_jointICC(quiz_object)
}
\arguments{
\item{quiz_object}{Receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the ICC
}
\description{
This function computes a graph of the ICC of the Rasch model
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)
plot_jointICC(quiz_object)
}
