% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_order.R
\name{plot_order}
\alias{plot_order}
\title{Order plot with ggplot2}
\usage{
plot_order(df_quizzes)
}
\arguments{
\item{df_quizzes}{Receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the order of answering
}
\description{
This function computes a graph with the order of answering of each student for each quiz.
}
\examples{
file_to_read <- "../../datasets/Dataset1/Quiz3_session12098.csv"
long_format <- read_lc(file_to_read)
plot_order(long_format)
}
