% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_personItem.R
\name{plot_personItem}
\alias{plot_personItem}
\title{Person - Item map}
\usage{
plot_personItem(quiz_object)
}
\arguments{
\item{quiz_object}{Receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the person-item parameters
}
\description{
This function computes a graph of the Person-Item map of the Rasch model
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)
plot_personItem(quiz_object)
}
