% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_personParameter.R
\name{plot_personParameter}
\alias{plot_personParameter}
\title{Plot of the person parameter of the Rasch model}
\usage{
plot_personParameter(quiz_object)
}
\arguments{
\item{quiz_object}{Receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the person parameter of the Rasch model
}
\description{
This function computes a graph of the person-parameter of the Rasch model
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)
plot_personParameter(quiz_object)
}
