% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_parse.R
\name{read_lc}
\alias{read_lc}
\title{Read file exported from Learning Catalytics or with the specified format.}
\usage{
read_lc(file)
}
\arguments{
\item{file}{The name of the csv file to read}
}
\value{
it returns a quiz object containing the following columns:  
"email address", "question", "responded at", "score" and "quiz"
}
\description{
This function receives as input a learning catalytics *csv* 
file and parse it to create quiz object.
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)
}
