% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_ETL.R
\name{plot_etl}
\alias{plot_etl}
\title{Easiness-time-cognitive level plot.}
\usage{
plot_etl(df_quizzes, challengeLevel)
}
\arguments{
\item{df_quizzes}{dataframe in long format of all quizzes (Generated with read_lc)}

\item{challengeLevel}{file of challengue level}
}
\value{
The ETL plot
}
\description{
This function computes a graph displaying the guessers plot for each quiz.
}
\examples{
file_to_read <- "../../datasets/Dataset1/Quiz3_session12098.csv"
file_cognitivelevel <- "../../datasets/Quiz2013-14_cognitive level_HB.csv" 
long_format <- read_lc(file_to_read)
cognitive_level <- data.frame(read_csv(file_cognitivelevel))
plot_etl(long_format, cognitive_level)
}
