% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_ETL.R
\name{plot_etl}
\alias{plot_etl}
\title{Easiness-time-level plot.}
\usage{
plot_etl(quiz_object, challengeLevel, item = "MCM.2014.item",
  rating = "Rating.HB")
}
\arguments{
\item{challengeLevel}{file of challengue level}

\item{df_quizzes}{dataframe in long format of all quizzes (Generated with read_lc)}
}
\value{
The ETL plot
}
\description{
This function computes a graph displaying the guessers plot for each quiz.
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
file_cognitivelevel <- "../../datasets/sample_dataset/cognitive_file.csv" 
quiz_object <- read_lc(file_to_read)
cognitive_level <- data.frame(read_csv(file_cognitivelevel))
quiz_object <- add_times(quiz_object)
plot_etl(quiz_object, cognitive_level)
}
