% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_plot_ind.R
\name{plot_group}
\alias{plot_group}
\title{Rasch model plot with ggplot2}
\usage{
plot_group(quiz_object, email_filtered, df_exam = NULL)
}
\arguments{
\item{quiz_object}{Receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the rasch model
}
\description{
This function computes a graph of the ICC of the Rasch model
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object_1 <- read_lc(file_to_read)
file_to_read <- "../../datasets/sample_dataset/Q02.csv"
quiz_object_2 <- read_lc(file_to_read)
file_to_read <- "../../datasets/sample_dataset/Q03.csv"
quiz_object_3 <- read_lc(file_to_read)
file_to_read <- "../../datasets/sample_dataset/Q04.csv"
quiz_object_4 <- read_lc(file_to_read)
quiz_object <- bind_rows(list(quiz_object_1, quiz_object_2, quiz_object_3, quiz_object_4))
final_exam <- "../../datasets/sample_dataset/finalexam_file.csv"
final_exam <- read_csv(final_exam)
email_filtered <- "UTKIJ1509E@gmail.com"
plot_group(quiz_object, email_filtered, final_exam)
}
