% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_guessers.R
\name{plot_guessers}
\alias{plot_guessers}
\title{Guessers per quiz plot}
\usage{
plot_guessers(quiz_object)
}
\arguments{
\item{df_quizzes}{receives as input the dataframe un long format of all quizzes (Generated with read_lc)}
}
\value{
A plot of the guessers of the selected quiz
}
\description{
This function computes a graph displaying the guessers plot for each quiz.
}
\examples{
file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)
quiz_object <- add_times(quiz_object)
plot_guessers(quiz_object)
}
