% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_parse.R
\name{read_lc}
\alias{read_lc}
\title{Read file exported from Learning Catalytics}
\usage{
read_lc(file)
}
\arguments{
\item{file}{The name of the csv file to read}
}
\value{
it returns a data frame containing the columns:
}
\description{
This function receives as input a learning catalytics *csv* 
file and parse it to create a long format data frame.
}
\examples{
file_to_read <- "../../datasets/Dataset1/Quiz3_session12098.csv"
head(read_lc(file_to_read))
}
